agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.21.0",
    "google-generativeai>=0.8.6",
    "ipython>=9.8.0",
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
